+++ doc/changes.md
@@ -545,6 +545,7 @@ as part of K95 at this time, the default terminal remains VT220 for now.
  - Fixed SGR font attributes not turning off if the reverse video attribute is 
    set (K95 bug 843)
  - Fixed cursor positioning on double-height/double-wide lines in K95G (K95 bug 844)
+ - Fixed ISO-2022 state not being restored when switching back from UTF-8 mode
 
 ## Kermit 95 v3.0 beta 7 - 27 January 2025